+++ lib/src/analyzers/lint_analyzer/metrics_records_builder.dart
@@ -4,7 +4,7 @@ import '../models/scoped_function_declaration.dart';
 import 'reporters/reports_builder.dart';
 
 /// [FileRecord] builder
-@Deprecated('will be removed in 3.3')
+@Deprecated('will be removed in 4.0')
 abstract class MetricsRecordsBuilder extends ReportsBuilder {
   void recordFunctionData(ScopedFunctionDeclaration declaration, Report record);
 }