+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.2.2
+
+* Update deprecation version for `Lines of Executable Code` `ConsoleReporter`, `MetricsAnalysisRunner`, `MetricsAnalyzer`, `MetricsRecordsBuilder` and `MetricsRecordsStore`.
+
 ## 3.2.1
 
 * Remove unnecessary scan by `Lines of Executable Code`